Feat:코드 리펙토링 및 gameClear 추가
