+++ src/componants/Game/Canvas.ts
@@ -1,0 +1,2 @@
+export const wrraperWidth = window.innerWidth;
+export const wrapperHeight = window.innerHeight;
