+++ src/componants/Game/Enemy.ts
@@ -1,31 +1,26 @@
-import BackGroundImg from '../img/spaceBackground.png';
-import spaceshiptImg from '../img/spaceBattleShip.png';
-import bullet from '../img/bullet.png';
-import enemy from '../img/meteor.png';
+import { wrapperHeight, wrraperWidth } from './Canvas';
 
-import type { Coordinate } from './Type';
+import type { Enemy } from './Type';
 
-let enemyList: Coordinate[] = [];
-
-const wrraperWidth = window.innerWidth;
-const wrapperHeight = window.innerHeight;
-let spaceShipX = wrraperWidth / 2 - 28;
-let spaceShipY = wrapperHeight - 60;
-
+let enemyList: Enemy[] = [];
+let gameOver = false;
 const createEnemy = () => {
   const interval = setInterval(() => {
     let max = wrraperWidth - 77;
     let randomNumber = Math.floor(Math.random() * (max + 1));
-    const newEnemy: Coordinate = {
+    const newEnemy: Enemy = {
       x: randomNumber, // 여기서 초기 위치 설정 필요
       y: 0,
       update: function () {
         this.x = randomNumber;
-        this.y += 7; // 적군 이동 업데이트
+        this.y += 4; // 적군 이동 업데이트
+        if (this.y >= wrapperHeight - 40) {
+          gameOver = true;
+          console.log('gameOver');
+        }
       },
     };
     enemyList.push(newEnemy);
-    console.log('적군 리스트', enemyList);
-  }, 30000);
+  }, 1000);
 };
-export { enemyList, createEnemy };
+export { enemyList, createEnemy, gameOver };
