+++ src/componants/Game/KeyCode.ts
@@ -1,10 +1,8 @@
 import { bulletList, createBullet } from './Bullet';
+import { wrapperHeight, wrraperWidth } from './Canvas';
 import { enemyList } from './Enemy';
 
 const keysDown: { [key: number]: boolean } = {};
-
-const wrraperWidth = window.innerWidth;
-const wrapperHeight = window.innerHeight;
 let spaceShipX = wrraperWidth / 2 - 28;
 let spaceShipY = wrapperHeight - 60;
 
@@ -22,16 +20,16 @@ const setUpKeyboardListener = () => {
 
 const update = () => {
   if (39 in keysDown) {
-    spaceShipX += 3;
+    spaceShipX += 5;
   }
   if (37 in keysDown) {
-    spaceShipX -= 3;
+    spaceShipX -= 5;
   }
   if (40 in keysDown) {
-    spaceShipY += 3;
+    spaceShipY += 5;
   }
   if (38 in keysDown) {
-    spaceShipY -= 3;
+    spaceShipY -= 5;
   }
   if (spaceShipX <= 0) {
     spaceShipX = 0;
@@ -47,11 +45,14 @@ const update = () => {
   }
   //총알의 y좌표 업데이트 하는 함수 호출
   for (let i = 0; i < bulletList.length; i++) {
-    bulletList[i].update();
+    if (bulletList[i].alive) {
+      bulletList[i].update();
+      bulletList[i].checkHit();
+    }
   }
   for (let i = 0; i < enemyList.length; i++) {
     enemyList[i].update();
   }
 };
 
-export { setUpKeyboardListener, update };
+export { spaceShipX, spaceShipY, setUpKeyboardListener, update };
