+++ src/componants/Game/Type.ts
@@ -1,5 +1,13 @@
-export type Coordinate = {
+export type Enemy = {
   x: number;
   y: number;
   update: () => void;
+};
+
+export type Bullet = {
+  x: number;
+  y: number;
+  alive: boolean;
+  update: () => void;
+  checkHit: () => void;
 };
