+++ src/shared/Router.tsx
@@ -7,7 +7,6 @@ import List from '../pages/List';
 import Calender from '../pages/Calender';
 import Quiz from '../pages/Quiz';
 import Game from '../pages/Game';
-import TestGame from '../pages/TestGame';
 
 function Router() {
   return (
@@ -20,7 +19,7 @@ function Router() {
           <Route path="/calender" element={<Calender />} />
           <Route path="/quiz" element={<Quiz />} />
         </Route>
-        <Route path="/game" element={<TestGame />} />
+        <Route path="/game" element={<Game />} />
       </Routes>
     </BrowserRouter>
   );
